pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evidence-capture-app"
version = "0.1.0"
description = "Manual testing evidence capture tool (PyQt6)"
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.7.0",
    "python-docx>=1.1.2",
    "Pillow>=10.4.0",
    "mss>=9.0.2",
    "pyyaml>=6.0.2",
]
